# src/user-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Hospital Microservice",
        "title": "Hospital Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3080",
    "basePath": "/user-service",
    "paths": {
...
        "/account/update": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update profile data for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update current user",
                "parameters": [
                    {
                        "description": "Update Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.UserUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResErr"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ResErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ResErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.UserUpdateReq": {
            "type": "object",
            "required": [
                "address",
                "avatar",
                "birthday",
                "email",
                "full_name",
                "gender"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                }
            }
        },
...
        "response.ResErr": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {your-token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}